# refactor: remove AIG optimizations