--- a/LeanSAT/AIG/Cached.lean
+++ b/LeanSAT/AIG/Cached.lean
@@ -74,40 +74,29 @@ def mkGateCached (aig : AIG α) (input : GateInput aig) : Entrypoint α :=
   let rhs := input.rhs.ref.gate
   let linv := input.lhs.inv
   let rinv := input.rhs.inv
-  have := input.lhs.ref.hgate
-  have := input.rhs.ref.hgate
+  have lhgate := input.lhs.ref.hgate
+  have rhgate := input.rhs.ref.hgate
   let decl := .gate lhs rhs linv rinv
   match cache.find? decl with
   | some hit =>
     ⟨⟨decls, cache, inv⟩, ⟨hit.idx, hit.hbound⟩⟩
   | none =>
     /-
-    Here we implement the constant propagating subset of:
+    We looked into implementing AIG optimization according to:
     https://fmv.jku.at/papers/BrummayerBiere-MEMICS06.pdf
-    TODO: rest of the table
+    However the constant folding subset of this seemed to *decrease* SAT solver performance noticeably.
+    We thus decided to just do the naive approach.
     -/
-    match decls[lhs], decls[rhs], linv, rinv with
-    -- Boundedness
-    | .const true, _, true, _ | .const false, _, false, _
-    | _, .const true, _, true | _, .const false, _, false =>
-      mkConstCached ⟨decls, cache, inv⟩ false
-    -- Left Neutrality
-    | .const true, _, false, false | .const false, _, true, false =>
-      ⟨⟨decls, cache, inv⟩, rhs, (by assumption)⟩
-    -- Right Neutrality
-    | _, .const true, false, false | _, .const false, false, true =>
-      ⟨⟨decls, cache, inv⟩, lhs, (by assumption)⟩
-    | _, _, _, _ =>
-      let g := decls.size
-      let cache := cache.insert decls decl
-      let decls := decls.push decl
-      have inv := by
-        intro lhs rhs linv rinv i h1 h2
-        simp only [decls] at *
-        simp only [Array.get_push] at h2
-        split at h2
-        . apply inv <;> assumption
-        . injections; omega
-      ⟨⟨decls, cache, inv⟩, ⟨g, by simp [g, decls]⟩⟩
+    let g := decls.size
+    let cache := cache.insert decls decl
+    let decls := decls.push decl
+    have inv := by
+      intro lhs rhs linv rinv i h1 h2
+      simp only [decls] at *
+      simp only [Array.get_push] at h2
+      split at h2
+      . apply inv <;> assumption
+      . injections; omega
+    ⟨⟨decls, cache, inv⟩, ⟨g, by simp [g, decls]⟩⟩
 
 end AIG