--- a/Test/AIG.lean
+++ b/Test/AIG.lean
@@ -1,3 +1,2 @@
-import Test.AIG.Opt
 import Test.AIG.Shared
 import Test.AIG.Stress