… Test/AIG/Opt.lean
… Test/Sat.lean-test-7-47.lrat
… Test/sat_check.lrat