@hargoniX 2024-06-04
Decided not to do this as in big problems constant propagation might actually be highly relevant, simply to reduce the size of problems.